Moved type dispatching related files to utilities/.
