+++ cpp/src/dataframe/cudf_table.cuh
@@ -29,7 +29,7 @@
 #include <thrust/execution_policy.h>
 #include <thrust/device_vector.h>
 #include <cassert>
-#include "types/type_dispatcher.hpp"
+#include "utilities/type_dispatcher.hpp"
 
 template <typename size_type>
 struct ValidRange {
