+++ cpp/src/hash/hash_functions.cuh
@@ -17,7 +17,7 @@
 #ifndef HASH_FUNCTIONS_CUH
 #define HASH_FUNCTIONS_CUH
 
-#include "types/wrapper_types.hpp"
+#include "utilities/wrapper_types.hpp"
 
 using hash_value_type = uint32_t;
