+++ cpp/src/utilities/wrapper_types.hpp
